+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 2.8.6, 4/5/2024
+Library Updates
+netty 4.1.108
+tomacat 9.0.86
+vertx-core 4.5.4
+
 ## Version 2.8.5, 2/9/2024
 
 ### Added